+++ Source/LightBlock/model/LightBlockModel.h
@@ -36,7 +36,7 @@ public:
 
 	virtual Array<WeakReference<Controllable>> getModelParameters() override;
 
-	virtual Array<Colour> getColors(int id, int resolution, float time, var params) override;
+	virtual Array<Colour> getColors(int id, int resolution, double time, var params) override;
 
 	void onControllableFeedbackUpdateInternal(ControllableContainer * cc, Controllable * c) override;
 	void childStructureChanged(ControllableContainer * cc) override;
