+++ Source/LightBlock/model/blocks/timeline/TimelineBlock.cpp
@@ -41,7 +41,7 @@ LightBlockModelUI * TimelineBlock::createUI()
 	return new TimelineBlockUI(this);
 }
 
-Array<Colour> TimelineBlock::getColors(int id, int resolution, float time, var params)
+Array<Colour> TimelineBlock::getColors(int id, int resolution, double time, var params)
 {
 	return sequence.getColors(id, resolution, time, params);
