+++ Source/Node/nodes/prop/PropNode.cpp
@@ -23,7 +23,7 @@ PropNode::~PropNode()
 {
 }
 
-Array<Colour> PropNode::getColors(int id, int resolution, float time, var params)
+Array<Colour> PropNode::getColors(int id, int resolution, double time, var params)
 {
 	Array<Colour> c = inColors->getColors(id, resolution, time, params);
 	if (c.size() == 0) c = ColorNode::getColors(id, resolution, time, params);
