+++ Source/Node/nodes/prop/PropNode.h
@@ -20,7 +20,7 @@ public:
 
 	ColorSlot * inColors;
 
-	Array<Colour> getColors(int id, int resolution, float time, var params) override;
+	Array<Colour> getColors(int id, int resolution, double time, var params) override;
 
 	String getTypeString() const override { return "Prop"; }
 	static PropNode * create(var params) { return new PropNode(params); }
