+++ Source/Node/nodes/solidcolor/SolidColorNode.cpp
@@ -22,7 +22,7 @@ SolidColorNode::~SolidColorNode()
 {
 }
 
-Array<Colour> SolidColorNode::getColors(int id, int resolution, float time, var params)
+Array<Colour> SolidColorNode::getColors(int id, int resolution, double time, var params)
 {
 	Array<Colour> result;
 	result.resize(resolution);
